apps/api/pyproject.toml
[tool.poetry]
name = "omnipresence"
version = "0.1.0"
description = "Unified presence and attendance management platform"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
requires-python = ">=3.11"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14"
drf-spectacular = "^0.27"
django-cors-headers = "^4.3"
mysqlclient = "^2.2"
django-extensions = "^3.2"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-django = "^4.7"
pytest-cov = "^4.1"
black = "^23.12"
isort = "^5.13"
flake8 = "^7.0"
mypy = "^1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line-length = 100

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
